backend/src/opmas/agents/example_agent/requirements.txt
# Inherit base agent requirements
-r ../base_agent_package/requirements.txt

# Additional dependencies for example agent
pytest-mock==3.12.0
